Crate tidext

Tidext is the official Tidechain client used by Tidefi, Quorums and Oracle.

Currently this client is geared toward subxt implementation.

Signer — Stronghold is aggressively feature gated so your client can use regular signer for testing purpose.

Stronghold signer

Stronghold is a secure software implementation with the sole purpose of isolating digital secrets from exposure to hackers and accidental leaks. It uses encrypted snapshots that can be easily backed up and securely shared between devices. Written in stable rust, it has strong guarantees of memory safety and process integrity.

Initialize stronghold signer from seed

use tidext::TidefiKeyring;

TidefiKeyring::try_from_seed("sr25519 raw seed or mnemonic", None)?;

Initialize stronghold signer from existing snapshot

use tidext::TidefiKeyring;

TidefiKeyring::try_from_stronghold_path("~/.stronghold", None, Some("stronghold password"))?;
   

Initialize stronghold signer from existing stronghold instance

use tidext::TidefiKeyring;
use iota_stronghold::Stronghold;

// This should run into a separate thread, see examples for more details
let stronghold = Stronghold::init_stronghold_system(vec![], vec![]);

TidefiKeyring::try_from_stronghold_instance(stronghold, None)?;
   

Tidext client

All functions are directly generated from tidechain metadata with subxt

Basic usage

use tidext::ClientBuilder;

// This needs to run on a different thread
let client = ClientBuilder::new().build().await?;
let swap_fee = client.swap_fee().await?;

Access the runtime API

Convert the client to a runtime API wrapper for custom runtime access

The subxt proc macro will provide methods to submit extrinsic(s) and read storage specific to the Tidechain

use tidext::ClientBuilder;

let client = ClientBuilder::new().build().await?;
let tides = tidext::with_runtime! {
  client,
  current_runtime,
  {
   client
    .runtime()
    .storage()
    .at(None)
    .await?
    .fetch(&current_runtime.storage().balances().total_issuance())
    .await?
   }
  }

Re-exports

pub use tidefi_primitives as primitives;

pub use iota_stronghold as stronghold;

Modules

lagoon

lagoon6030

lagoon6040

lagoon7000

test_utils

Test utils

tidechain

tidechain6030

tidechain6040

tidechain7000

Macros

__with_tidext_runtime

latest_block

Get latest Tidechain block

make_rpc_call

Make RPC call to Tidechain

Structs

Client

Tidechain client

ClientBuilder

Tidechain Client builder

Lagoon6030Runtime

Lagoon6040Runtime

Lagoon7000Runtime

NodeHealth

Health status of the node the client is currently connected

Permill

A fixed point representation of a number in the range [0, 1].

StrongholdSigner

Stronghold signer instance.

Tidechain6030Runtime

Tidechain6040Runtime

Tidechain7000Runtime

TidechainConfig

Custom Config implementation with Clone

TidextKeyring

Tidefi keyring backed with a stronghold pair signer.

Enums

AssetsCall

BalancesCall

Error

MultiAddress

A multi-format address wrapper for on-chain accounts.

OracleCall

QuorumCall

RewardDestination

A destination account for payment

StakingCall

SubstrateSubxtError

The underlying error enum, generic over the type held by the Runtime variant. Prefer to use the Error<E> and Error aliases over using this type directly.

TidechainCall

TidefiCall

TidefiRuntime

Traits

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

OracleExt

An extension trait for Client that provides a variety of convenient Oracle functions.

QuorumExt

An extension trait for Client that provides a variety of convenient Quorum functions.

Signer

Signing transactions requires a Signer. This is responsible for providing the “from” account that the transaction is being signed by, as well as actually signing a SCALE encoded payload.

Functions

get_signer_from_stronghold

Try to get signer details for an existing stronghold instance at the specific location

init_ephemeral_stronghold

Initialize the ephemeral stronghold client with a seed.

init_stronghold_from_path

Initialize a new stronghold instance from the provided snapshot path and passphrase

init_stronghold_from_seed

Initialize a new stronghold instance from the sr25519 mnemonic or raw seed

Type Definitions

KeyPair

Tidechain key-pair representation.

TidefiKeyring

Tidefi keyring

TidefiPairSigner

Stronghold pair signer.

Derive Macros

Decode

Derive parity_scale_codec::Decode and for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.