Struct tidext::lagoon6040::staking::storage::StorageApi

pub struct StorageApi;

Implementations

impl StorageApi

pub fn history_depth(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, Yes, ()>

Number of eras to keep in history.

Information is kept for eras in [current_era - history_depth; current_era].

Must be more than the number of eras delayed by session otherwise. I.e. active era must always be in history. I.e. active_era > current_era - history_depth must be guaranteed.

pub fn validator_count(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, Yes, ()>

The ideal number of staking participants.

pub fn minimum_validator_count(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, Yes, ()>

Minimum number of staking participants before emergency conditions are imposed.

pub fn invulnerables(
&self
) -> StaticStorageAddress<DecodeStaticType<Vec<AccountId32>>, Yes, Yes, ()>

Any validators that may never be slashed or forcibly kicked. It’s a Vec since they’re easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

pub fn bonded(
&self,
_0: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<AccountId32>, Yes, (), Yes>

Map from all locked “stash” accounts to the controller account.

pub fn bonded_root(
&self
) -> StaticStorageAddress<DecodeStaticType<AccountId32>, (), (), Yes>

Map from all locked “stash” accounts to the controller account.

pub fn min_nominator_bond(
&self
) -> StaticStorageAddress<DecodeStaticType<u128>, Yes, Yes, ()>

The minimum active bond to become and maintain the role of a nominator.

pub fn min_validator_bond(
&self
) -> StaticStorageAddress<DecodeStaticType<u128>, Yes, Yes, ()>

The minimum active bond to become and maintain the role of a validator.

pub fn min_commission(
&self
) -> StaticStorageAddress<DecodeStaticType<Perbill>, Yes, Yes, ()>

The minimum amount of commission that validators can set.

If set to 0, no limit exists.

pub fn ledger(
&self,
_0: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<StakingLedger>, Yes, (), Yes>

Map from all (unlocked) “controller” accounts to the info regarding the staking.

pub fn ledger_root(
&self
) -> StaticStorageAddress<DecodeStaticType<StakingLedger>, (), (), Yes>

Map from all (unlocked) “controller” accounts to the info regarding the staking.

pub fn payee(
&self,
_0: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<RewardDestination<AccountId32>>, Yes, Yes, Yes>

Where the reward payment should be made. Keyed by stash.

pub fn payee_root(
&self
) -> StaticStorageAddress<DecodeStaticType<RewardDestination<AccountId32>>, (), Yes, Yes>

Where the reward payment should be made. Keyed by stash.

pub fn validators(
&self,
_0: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<ValidatorPrefs>, Yes, Yes, Yes>

The map from (wannabe) validator stash key to the preferences of that validator.

pub fn validators_root(
&self
) -> StaticStorageAddress<DecodeStaticType<ValidatorPrefs>, (), Yes, Yes>

The map from (wannabe) validator stash key to the preferences of that validator.

pub fn counter_for_validators(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, Yes, ()>

Counter for the related counted storage map

pub fn max_validators_count(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, (), ()>

The maximum validator count before we stop allowing new validators to join.

When this value is not set, no limits are enforced.

pub fn nominators(
&self,
_0: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<Nominations>, Yes, (), Yes>

The map from nominator stash key to their nomination preferences, namely the validators that they wish to support.

Note that the keys of this storage map might become non-decodable in case the [Config::MaxNominations] configuration is decreased. In this rare case, these nominators are still existent in storage, their key is correct and retrievable (i.e. contains_key indicates that they exist), but their value cannot be decoded. Therefore, the non-decodable nominators will effectively not-exist, until they re-submit their preferences such that it is within the bounds of the newly set Config::MaxNominations.

This implies that ::iter_keys().count() and ::iter().count() might return different values for this map. Moreover, the main ::count() is aligned with the former, namely the number of keys that exist.

Lastly, if any of the nominators become non-decodable, they can be chilled immediately via [Call::chill_other] dispatchable by anyone.

pub fn nominators_root(
&self
) -> StaticStorageAddress<DecodeStaticType<Nominations>, (), (), Yes>

The map from nominator stash key to their nomination preferences, namely the validators that they wish to support.

Note that the keys of this storage map might become non-decodable in case the [Config::MaxNominations] configuration is decreased. In this rare case, these nominators are still existent in storage, their key is correct and retrievable (i.e. contains_key indicates that they exist), but their value cannot be decoded. Therefore, the non-decodable nominators will effectively not-exist, until they re-submit their preferences such that it is within the bounds of the newly set Config::MaxNominations.

This implies that ::iter_keys().count() and ::iter().count() might return different values for this map. Moreover, the main ::count() is aligned with the former, namely the number of keys that exist.

Lastly, if any of the nominators become non-decodable, they can be chilled immediately via [Call::chill_other] dispatchable by anyone.

pub fn counter_for_nominators(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, Yes, ()>

Counter for the related counted storage map

pub fn max_nominators_count(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, (), ()>

The maximum nominator count before we stop allowing new validators to join.

When this value is not set, no limits are enforced.

pub fn current_era(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, (), ()>

The current era index.

This is the latest planned era, depending on how the Session pallet queues the validator set, it might be active or not.

pub fn active_era(
&self
) -> StaticStorageAddress<DecodeStaticType<ActiveEraInfo>, Yes, (), ()>

The active era information, it holds index and start.

The active era is the era being currently rewarded. Validator set of this era must be equal to [SessionInterface::validators].

pub fn eras_start_session_index(
&self,
_0: impl Borrow<u32>
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, (), Yes>

The session index at which the era start for the last HISTORY_DEPTH eras.

Note: This tracks the starting session (i.e. session index when era start being active) for the eras in [CurrentEra - HISTORY_DEPTH, CurrentEra].

pub fn eras_start_session_index_root(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, (), (), Yes>

The session index at which the era start for the last HISTORY_DEPTH eras.

Note: This tracks the starting session (i.e. session index when era start being active) for the eras in [CurrentEra - HISTORY_DEPTH, CurrentEra].

pub fn eras_stakers(
&self,
_0: impl Borrow<u32>,
_1: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<Exposure<AccountId32, u128>>, Yes, Yes, Yes>

Exposure of validator at era.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

pub fn eras_stakers_root(
&self
) -> StaticStorageAddress<DecodeStaticType<Exposure<AccountId32, u128>>, (), Yes, Yes>

Exposure of validator at era.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

pub fn eras_stakers_clipped(
&self,
_0: impl Borrow<u32>,
_1: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<Exposure<AccountId32, u128>>, Yes, Yes, Yes>

Clipped Exposure of validator at era.

This is similar to [ErasStakers] but number of nominators exposed is reduced to the T::MaxNominatorRewardedPerValidator biggest stakers. (Note: the field total and own of the exposure remains unchanged). This is used to limit the i/o cost for the nominator payout.

This is keyed fist by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

pub fn eras_stakers_clipped_root(
&self
) -> StaticStorageAddress<DecodeStaticType<Exposure<AccountId32, u128>>, (), Yes, Yes>

Clipped Exposure of validator at era.

This is similar to [ErasStakers] but number of nominators exposed is reduced to the T::MaxNominatorRewardedPerValidator biggest stakers. (Note: the field total and own of the exposure remains unchanged). This is used to limit the i/o cost for the nominator payout.

This is keyed fist by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

pub fn eras_validator_prefs(
&self,
_0: impl Borrow<u32>,
_1: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<ValidatorPrefs>, Yes, Yes, Yes>

Similar to ErasStakers, this holds the preferences of validators.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras.

pub fn eras_validator_prefs_root(
&self
) -> StaticStorageAddress<DecodeStaticType<ValidatorPrefs>, (), Yes, Yes>

Similar to ErasStakers, this holds the preferences of validators.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras.

pub fn eras_validator_reward(
&self,
_0: impl Borrow<u32>
) -> StaticStorageAddress<DecodeStaticType<u128>, Yes, (), Yes>

The total validator era payout for the last HISTORY_DEPTH eras.

Eras that haven’t finished yet or has been removed doesn’t have reward.

pub fn eras_validator_reward_root(
&self
) -> StaticStorageAddress<DecodeStaticType<u128>, (), (), Yes>

The total validator era payout for the last HISTORY_DEPTH eras.

Eras that haven’t finished yet or has been removed doesn’t have reward.

pub fn eras_reward_points(
&self,
_0: impl Borrow<u32>
) -> StaticStorageAddress<DecodeStaticType<EraRewardPoints<AccountId32>>, Yes, Yes, Yes>

Rewards for the last HISTORY_DEPTH eras. If reward hasn’t been set or has been removed then 0 reward is returned.

pub fn eras_reward_points_root(
&self
) -> StaticStorageAddress<DecodeStaticType<EraRewardPoints<AccountId32>>, (), Yes, Yes>

Rewards for the last HISTORY_DEPTH eras. If reward hasn’t been set or has been removed then 0 reward is returned.

pub fn eras_total_stake(
&self,
_0: impl Borrow<u32>
) -> StaticStorageAddress<DecodeStaticType<u128>, Yes, Yes, Yes>

The total amount staked for the last HISTORY_DEPTH eras. If total hasn’t been set or has been removed then 0 stake is returned.

pub fn eras_total_stake_root(
&self
) -> StaticStorageAddress<DecodeStaticType<u128>, (), Yes, Yes>

The total amount staked for the last HISTORY_DEPTH eras. If total hasn’t been set or has been removed then 0 stake is returned.

pub fn force_era(
&self
) -> StaticStorageAddress<DecodeStaticType<Forcing>, Yes, Yes, ()>

Mode of era forcing.

pub fn slash_reward_fraction(
&self
) -> StaticStorageAddress<DecodeStaticType<Perbill>, Yes, Yes, ()>

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

pub fn canceled_slash_payout(
&self
) -> StaticStorageAddress<DecodeStaticType<u128>, Yes, Yes, ()>

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

pub fn unapplied_slashes(
&self,
_0: impl Borrow<u32>
) -> StaticStorageAddress<DecodeStaticType<Vec<UnappliedSlash<AccountId32, u128>>>, Yes, Yes, Yes>

All unapplied slashes that are queued for later.

pub fn unapplied_slashes_root(
&self
) -> StaticStorageAddress<DecodeStaticType<Vec<UnappliedSlash<AccountId32, u128>>>, (), Yes, Yes>

All unapplied slashes that are queued for later.

pub fn bonded_eras(
&self
) -> StaticStorageAddress<DecodeStaticType<Vec<(u32, u32)>>, Yes, Yes, ()>

A mapping from still-bonded eras to the first session index of that era.

Must contains information for eras for the range: [active_era - bounding_duration; active_era]

pub fn validator_slash_in_era(
&self,
_0: impl Borrow<u32>,
_1: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<(Perbill, u128)>, Yes, (), Yes>

All slashing events on validators, mapped by era to the highest slash proportion and slash value of the era.

pub fn validator_slash_in_era_root(
&self
) -> StaticStorageAddress<DecodeStaticType<(Perbill, u128)>, (), (), Yes>

All slashing events on validators, mapped by era to the highest slash proportion and slash value of the era.

pub fn nominator_slash_in_era(
&self,
_0: impl Borrow<u32>,
_1: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<u128>, Yes, (), Yes>

All slashing events on nominators, mapped by era to the highest slash value of the era.

pub fn nominator_slash_in_era_root(
&self
) -> StaticStorageAddress<DecodeStaticType<u128>, (), (), Yes>

All slashing events on nominators, mapped by era to the highest slash value of the era.

pub fn slashing_spans(
&self,
_0: impl Borrow<AccountId32>
) -> StaticStorageAddress<DecodeStaticType<SlashingSpans>, Yes, (), Yes>

Slashing spans for stash accounts.

pub fn slashing_spans_root(
&self
) -> StaticStorageAddress<DecodeStaticType<SlashingSpans>, (), (), Yes>

Slashing spans for stash accounts.

pub fn span_slash(
&self,
_0: impl Borrow<AccountId32>,
_1: impl Borrow<u32>
) -> StaticStorageAddress<DecodeStaticType<SpanRecord<u128>>, Yes, Yes, Yes>

Records information about the maximum slash of a stash within a slashing span, as well as how much reward has been paid out.

pub fn span_slash_root(
&self
) -> StaticStorageAddress<DecodeStaticType<SpanRecord<u128>>, (), Yes, Yes>

Records information about the maximum slash of a stash within a slashing span, as well as how much reward has been paid out.

pub fn earliest_unapplied_slash(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, (), ()>

The earliest era for which we have a pending, unapplied slash.

pub fn current_planned_session(
&self
) -> StaticStorageAddress<DecodeStaticType<u32>, Yes, Yes, ()>

The last planned session scheduled by the session pallet.

This is basically in sync with the call to [pallet_session::SessionManager::new_session].

pub fn offending_validators(
&self
) -> StaticStorageAddress<DecodeStaticType<Vec<(u32, bool)>>, Yes, Yes, ()>

Indices of validators that have offended in the active era and whether they are currently disabled.

This value should be a superset of disabled validators since not all offences lead to the validator being disabled (if there was no slash). This is needed to track the percentage of validators that have offended in the current era, ensuring a new era is forced if OffendingValidatorsThreshold is reached. The vec is always kept sorted so that we can find whether a given validator has previously offended using binary search. It gets cleared when the era ends.

pub fn storage_version(
&self
) -> StaticStorageAddress<DecodeStaticType<Releases>, Yes, Yes, ()>

True if network has been upgraded to this version. Storage version of the pallet.

This is set to v7.0.0 for new networks.

pub fn chill_threshold(
&self
) -> StaticStorageAddress<DecodeStaticType<Percent>, Yes, (), ()>

The threshold for when users can start calling chill_other for other validators / nominators. The threshold is compared to the actual number of validators / nominators (CountFor*) in the system compared to the configured max (Max*Count).