Struct tidext::Client

pub struct Client {
    pub gaz_limit: u64,
    pub signer: TidefiPairSigner,
    pub rpc_url: String,
    pub other_signatories: Vec<AccountId>,
    // some fields omitted
}

Tidechain client

Fields
gaz_limit: u64signer: TidefiPairSignerrpc_url: Stringother_signatories: Vec<AccountId>
Implementations

impl Client

pub fn set_signer(&mut self, signer: TidefiPairSigner)

Set new signer for the client.

pub fn runtime(&self) -> TidechainRuntimeApi

Subxt Runtime (Isolated, but not clonable)

pub fn get_account_id(&self) -> AccountId

Get account id for current signer.

pub async fn get_regular_swap_fee(&self) -> Result<f32, Error>

Get the swap percentage fee registered on-chain for a regular user

pub async fn get_market_maker_swap_fee(&self) -> Result<f32, Error>

Get the swap percentage fee registered on-chain for a market maker

pub async fn system_health(&self) -> Result<NodeHealth, Error>

Validate node connection. You should call this function every ~ 10 seconds to keep the connection alive

pub async fn submit_signed_extrinsic(
&self,
extrinsic: String
) -> Result<Hash, Error>

Submit extrinsic

pub async fn extrinsic_cost(&self, extrinsic: String) -> Result<Balance, Error>

Get the cost (gas fee) of an extrinsic on-chain (always in TIDE).

pub async fn quorum_status(&self) -> Result<bool, Error>

Current quorum status.

pub async fn total_stake_for(
&self,
currency_id: CurrencyId
) -> Result<Balance, Error>

Return the total staked for the currency.

pub async fn total_supply_for(
&self,
currency_id: CurrencyId
) -> Result<Balance, Error>

Return the total supply for the currency across all accounts.

Wrapped token are minted when deposited and burned when withdrawed, the total supply represent the total tokens deposited in chain and not withdrawed.

Total issuance for TiDE token, represent the number of tokens issued by the runtime.

pub async fn balance(
&self,
account_id: &AccountId,
currency_id: CurrencyId
) -> Result<CurrencyBalance<Balance>, Error>

Return available balance.

pub async fn balances(
&self,
account_id: &AccountId
) -> Result<Vec<(CurrencyId, CurrencyBalance<Balance>)>, Error>

Return available balances with all owned assets.

pub async fn all_assets(
&self
) -> Result<Vec<(CurrencyId, CurrencyMetadata<Vec<u8>>)>, Error>

Return a list of all assets registered on-chain.

pub async fn stakes(
&self,
account_id: &AccountId,
currency_id: Option<CurrencyId>
) -> Result<Vec<(CurrencyId, Stake<Balance, BlockNumber>)>, Error>

Return a list of all stakes for the AccountId with optional CurrencyId.

pub async fn withdrawal(
&self,
currency_id: CurrencyId,
amount: Balance,
external_address: Vec<u8>
) -> Result<(), Error>

Request withdrawal for the current signer.

pub async fn withdrawal_and_watch(
&self,
currency_id: CurrencyId,
amount: Balance,
external_address: Vec<u8>
) -> Result<(), Error>

Request withdrawal for the current signer and wait the processing in block.

pub async fn withdrawal_extrinsic(
&self,
asset_id: CurrencyId,
amount: Balance,
external_address: Vec<u8>
) -> Result<String, Error>

Initialize a withdrawal for the current signer.

pub async fn swap(
&self,
currency_id_from: CurrencyId,
amount_from: Balance,
currency_id_to: CurrencyId,
amount_to: Balance,
swap_type: SwapType,
slippage_tolerance: Option<Permill>
) -> Result<(), Error>

Initialize a swap for the current signer. The amount_from of the currency_id_from + the network fee eg. 2%, will be held untill the swap is processed or cancelled.

pub async fn cancel_swap(&self, request_id: Hash) -> Result<(), Error>

Cancel a swap and release the funds.

pub async fn swap_and_watch(
&self,
currency_id_from: CurrencyId,
amount_from: Balance,
currency_id_to: CurrencyId,
amount_to: Balance,
swap_type: SwapType,
slippage_tolerance: Option<Permill>
) -> Result<(), Error>

Initialize a swap for the current signer and wait the processing in block. The amount_from of the currency_id_from will be held untill the swap is processed or cancelled.

pub async fn swap_extrinsic(
&self,
currency_id_from: CurrencyId,
amount_from: Balance,
currency_id_to: CurrencyId,
amount_to: Balance,
swap_type: SwapType,
slippage_tolerance: Option<Permill>
) -> Result<String, Error>

Generate signed swap extrinsic for the current signer. The amount_from of the currency_id_from will be held untill the swap is processed or cancelled.

pub async fn stake(
&self,
currency_id: CurrencyId,
amount: Balance,
duration: u32
) -> Result<(), Error>

Request stake for the current signer.

pub async fn stake_and_watch(
&self,
currency_id: CurrencyId,
amount: Balance,
duration: u32
) -> Result<(), Error>

Request stake for the current signer and wait the processing in block.

pub async fn stake_extrinsic(
&self,
currency_id: CurrencyId,
amount: Balance,
duration: u32
) -> Result<String, Error>

Request stake extrinsic for the current signer.

pub async fn unstake(
&self,
stake_id: Hash,
force_unstake: bool
) -> Result<(), Error>

Request unstake for the current signer.

pub async fn unstake_and_watch(
&self,
stake_id: Hash,
force_unstake: bool
) -> Result<(), Error>

Request unstake for the current signer and wait the processing in block.

pub async fn unstake_extrinsic(
&self,
stake_id: Hash,
force_unstake: bool
) -> Result<String, Error>

Unstake extrinsic for the current signer.

pub async fn transfer(
&self,
destination_id: &AccountId,
currency_id: CurrencyId,
amount: Balance
) -> Result<(), Error>

Transfer currency to another account id from the current signer.

pub async fn transfer_and_watch(
&self,
destination_id: &AccountId,
currency_id: CurrencyId,
amount: Balance
) -> Result<(), Error>

Transfer currency to another account id from the current signer and wait the processing in block.

pub async fn transfer_extrinsic(
&self,
destination_id: &AccountId,
currency_id: CurrencyId,
amount: Balance
) -> Result<String, Error>

Transfer extrinsic for the current signer.

pub async fn submit_batch(&self, calls: Vec<TidechainCall>) -> Result<(), Error>

Submit batch call.