Module tidext::tidechain

Modules

asset_registry

assets

authority_discovery

authorship

babe

bags_list

balances

bounties

council

democracy

election_provider_multi_phase

elections

fees

grandpa

historical

identity

im_online

indices

multisig

offences

oracle

preimage

proxy

quorum

recovery

runtime_types

scheduler

security

session

staking

sudo

sunrise

system

technical_committee

technical_membership

tidefi

tidefi_staking

timestamp

transaction_payment

treasury

utility

vesting

Structs

ConstantsApi

StorageApi

TransactionApi

Enums

Event

Statics

PALLETS

Functions

constants

storage

tx

validate_codegen

check whether the Client you are using is aligned with the statically generated codegen.

Type Definitions

DispatchError

The default error type returned when there is a runtime issue, exposed here for ease of use.