Module tidext::tidechain::quorum::events

Structs

AccountChanged

Quorum account changed

BurnedAcknowledged

Quorum member acknowledged the burned and initiated the process

BurnedInitialized

Quorum burned token to the account from tidechain

ConfigurationUpdated

The quorum configuration has been updated, all elected members should re-submit public keys

Minted

Quorum minted token to the account

ProposalApproved

Proposal has been approved

ProposalProcessed

Proposal has been processed successfully

ProposalRejected

Proposal has been rejected

ProposalSubmitted

Proposal has been processed successfully

StatusChanged

Quorum status changed

VoteAgainst

Member voted against a proposal

VoteFor

Member voted for a proposal

WatchTransactionAdded

A new transaction has been added to the watch list