Trait tidext::OracleExt

pub trait OracleExt {
    fn update_status<'life0, 'async_trait>(
        &'life0 self, 
        enabled: bool
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_account_id<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        account_id: &'life1 AccountId
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn add_market_maker<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        account_id: &'life1 AccountId
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn remove_market_maker<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        account_id: &'life1 AccountId
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn confirm_swap<'life0, 'async_trait>(
        &'life0 self, 
        request_id: Hash, 
        market_makers: Vec<SwapConfirmation>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn cancel_swap<'life0, 'async_trait>(
        &'life0 self, 
        request_id: Hash
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn update_assets_value<'life0, 'async_trait>(
        &'life0 self, 
        values: Vec<(AssetId, Balance)>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

An extension trait for Client that provides a variety of convenient Oracle functions.

Required methods

fn update_status<'life0, 'async_trait>(
&'life0 self,
enabled: bool
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Update oracle status

fn update_account_id<'life0, 'life1, 'async_trait>(
&'life0 self,
account_id: &'life1 AccountId
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Update oracle account id

fn add_market_maker<'life0, 'life1, 'async_trait>(
&'life0 self,
account_id: &'life1 AccountId
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Add account id as a new trusted market maker

fn remove_market_maker<'life0, 'life1, 'async_trait>(
&'life0 self,
account_id: &'life1 AccountId
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Remove account id from the trusted market maker list

fn confirm_swap<'life0, 'async_trait>(
&'life0 self,
request_id: Hash,
market_makers: Vec<SwapConfirmation>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Confirm swap

fn cancel_swap<'life0, 'async_trait>(
&'life0 self,
request_id: Hash
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Cancel swap

fn update_assets_value<'life0, 'async_trait>(
&'life0 self,
values: Vec<(AssetId, Balance)>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Update assets value for the sunrise pool It should represent how many TDFY’s for 1 AssetId.

The value should always be formatted with TDFY decimals (12)

Example:

If the Bitcoin price is 0.001815 BTC (for 1 TDFY) You get 550.9641873278 TDFY for 1 BTC

The value should be: vec![(2, 550_964_187_327_800)]

If the ETH price is 0.03133 ETH (for 1 TDFY) You get 31.9182891796999 TDFY for 1 ETH

The value sent should be: vec![(4, 31_918_289_179_699)]

If the USDT price is 33.650000 USDT (for 1 TDFY) You get 0.029717682000 TDFY for 1 USDT

The value sent should be: vec![(4, 29_717_682_020)]