Trait tidext::QuorumExt

pub trait QuorumExt {
    fn get_proposals<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(Hash, BlockNumber, ProposalType<AccountId, BlockNumber, Vec<u8>, Vec<AccountId>>)>, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_burned_queue<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(Hash, Withdrawal<AccountId, BlockNumber, Vec<u8>>)>, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_proposal<'life0, 'async_trait>(
        &'life0 self,
        proposal: Hash
    ) -> Pin<Box<dyn Future<Output = Result<Option<ProposalVotes<BlockNumber, Vec<AccountId>>>, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn submit_proposal<'life0, 'async_trait>(
        &'life0 self,
        proposal: ProposalType<AccountId, BlockNumber, Vec<u8>, Vec<AccountId>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn acknowledge_proposal<'life0, 'async_trait>(
        &'life0 self,
        proposal: Hash
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn acknowledge_burned<'life0, 'async_trait>(
        &'life0 self,
        proposal: Hash
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn reject_proposal<'life0, 'async_trait>(
        &'life0 self,
        proposal: Hash
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn submit_public_keys<'life0, 'async_trait>(
        &'life0 self,
        public_keys: Vec<(AssetId, Vec<u8>)>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn is_quorum_member<'life0, 'life1, 'async_trait>(
        &'life0 self,
        account_id: &'life1 AccountId
    ) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

An extension trait for Client that provides a variety of convenient Quorum functions.

Required Methods

fn get_proposals<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<(Hash, BlockNumber, ProposalType<AccountId, BlockNumber, Vec<u8>, Vec<AccountId>>)>, Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Get all proposals return Vec<(proposal_id, initial_block, proposal_type)>

fn get_burned_queue<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Vec<(Hash, Withdrawal<AccountId, BlockNumber, Vec<u8>>)>, Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Get all pending burned

fn get_proposal<'life0, 'async_trait>(
&'life0 self,
proposal: Hash
) -> Pin<Box<dyn Future<Output = Result<Option<ProposalVotes<BlockNumber, Vec<AccountId>>>, Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Get proposal votes and status

fn submit_proposal<'life0, 'async_trait>(
&'life0 self,
proposal: ProposalType<AccountId, BlockNumber, Vec<u8>, Vec<AccountId>>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Submit new proposal

fn acknowledge_proposal<'life0, 'async_trait>(
&'life0 self,
proposal: Hash
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Acknowledge a proposal

fn acknowledge_burned<'life0, 'async_trait>(
&'life0 self,
proposal: Hash
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Acknowledge a burned item, and remove it from the queue

fn reject_proposal<'life0, 'async_trait>(
&'life0 self,
proposal: Hash
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Reject a proposal

fn submit_public_keys<'life0, 'async_trait>(
&'life0 self,
public_keys: Vec<(AssetId, Vec<u8>)>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Submit public keys for the current signer (required) for each members and wait finalization on-chain

You need to submit all public keys at once as all previous keys will be deleted for this account.

When a new configuration is set, all members need to set their public keys before they can submit, acknowledge and reject any proposal.

fn is_quorum_member<'life0, 'life1, 'async_trait>(
&'life0 self,
account_id: &'life1 AccountId
) -> Pin<Box<dyn Future<Output = Result<bool, Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Validate if an account id is currently part of the quorum set