Trait tidext::Signer

pub trait Signer<T> where
    T: Config, {
    fn nonce(&self) -> Option<<T as Config>::Index>;
    fn account_id(&self) -> &<T as Config>::AccountId;
    fn address(&self) -> <T as Config>::Address;
    fn sign(&self, signer_payload: &[u8]) -> <T as Config>::Signature;
}

Signing transactions requires a Signer. This is responsible for providing the “from” account that the transaction is being signed by, as well as actually signing a SCALE encoded payload. Optionally, a signer can also provide the nonce for the transaction to use.

Required methods

fn nonce(&self) -> Option<<T as Config>::Index>

Optionally returns a nonce.

fn account_id(&self) -> &<T as Config>::AccountId

Return the “from” account ID.

fn address(&self) -> <T as Config>::Address

Return the “from” address.

fn sign(&self, signer_payload: &[u8]) -> <T as Config>::Signature

Takes a signer payload for an extrinsic, and returns a signature based on it.

Some signers may fail, for instance because the hardware on which the keys are located has refused the operation.