Crate tidext

Re-exports

pub use crate::tidechain::runtime_types::lagoon_runtime::Call as TidechainCall;

pub use crate::tidechain::runtime_types::pallet_oracle::pallet::Call as OracleCall;

pub use crate::tidechain::runtime_types::pallet_quorum::pallet::Call as QuorumCall;

pub use crate::tidechain::runtime_types::pallet_tidefi::pallet::Call as TidefiCall;

pub use crate::tidechain::runtime_types::pallet_tidefi_stake::pallet::Call as TidefiStakingCall;

Modules

assets

error

keyring

test_utils

tidechain

Structs

Client

Tidechain client

ClientBuilder

NodeHealth

Health struct returned by the RPC

Permill

A fixed point representation of a number in the range [0, 1].

TidechainConfig

Custom Config impl with Clone

Enums

CurrencyId

Enum indicating the currency. Tide is the native token.

SwapType

Swap type

Traits

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

OracleExt

An extension trait for Client that provides a variety of convenient Oracle functions.

QuorumExt

An extension trait for Client that provides a variety of convenient Quorum functions.

Signer

Extrinsic signer.

Type Definitions

Balance

Balance of an account.

Hash

A hash of some data used by the chain.

SubstrateSubxtError

An error that may contain some runtime error E

TidechainRuntimeApi

Derive Macros

Decode

Derive parity_scale_codec::Decode and for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.