Crate tidext

Tidext is the official Tidechain client used by Tidefi, Quorums and Oracle.

Currently this client is geared toward subxt implementation.

Signer — Stronghold is aggressively feature gated so your client can use regular signer for testing purpose.

Stronghold signer

Stronghold is a secure software implementation with the sole purpose of isolating digital secrets from exposure to hackers and accidental leaks. It uses encrypted snapshots that can be easily backed up and securely shared between devices. Written in stable rust, it has strong guarantees of memory safety and process integrity.

Initialize stronghold signer from seed

use tidext::TidefiKeyring;

TidefiKeyring::try_from_seed("sr25519 raw seed or mnemonic", None)?;

Initialize stronghold signer from existing snapshot

use tidext::TidefiKeyring;

TidefiKeyring::try_from_stronghold_path("~/.stronghold", None, Some("stronghold password"))?;
   

Initialize stronghold signer from existing stronghold instance

use tidext::TidefiKeyring;
use iota_stronghold::Stronghold;

// This should run into a separate thread, see examples for more details
let stronghold = Stronghold::init_stronghold_system(vec![], vec![]);

TidefiKeyring::try_from_stronghold_instance(stronghold, None)?;
   

Tidext client

All functions are directly generated from tidechain metadata with subxt

Basic usage

use tidext::ClientBuilder;

// This needs to run on a different thread
let client = ClientBuilder::new().set_signer(signer).build().await?;
let swap_fee = client.swap_fee().await?;

Access the runtime API

Convert the client to a runtime api wrapper for custom runtime access

The subxt proc macro will provide methods to submit extrinsics and read storage specific to the Tidechain runtime

use tidext::ClientBuilder;

let client = ClientBuilder::new().set_signer(signer).build().await?;
let runtime = client.runtime();
let tides = runtime.storage().balances().total_issuance(None).await?;

Re-exports

pub use iota_stronghold as stronghold;

pub use tidefi_primitives as primitives;

Modules

test_utils

Test utils

tidechain

Macros

latest_block

Get latest Tidechain block

make_rpc_call

Make RPC call to Tidechain

query_storage

Query Tidechain storage

Structs

Client

Tidechain client

ClientBuilder

Tidechain Client builder

NodeHealth

Health status of the node the client is currently connected

Permill

A fixed point representation of a number in the range [0, 1].

StrongholdSigner

Stronghold signer instance.

TidechainConfig

Custom Config implementation with Clone

TidextKeyring

Tidefi keyring backed with a stronghold pair signer.

Enums

Error

OracleCall

Contains one variant per dispatchable that can be called by an extrinsic.

QuorumCall

Contains one variant per dispatchable that can be called by an extrinsic.

TidechainCall

TidefiCall

Contains one variant per dispatchable that can be called by an extrinsic.

TidefiStakingCall

Contains one variant per dispatchable that can be called by an extrinsic.

Traits

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

OracleExt

An extension trait for Client that provides a variety of convenient Oracle functions.

QuorumExt

An extension trait for Client that provides a variety of convenient Quorum functions.

Signer

Signing transactions requires a Signer. This is responsible for providing the “from” account that the transaction is being signed by, as well as actually signing a SCALE encoded payload. Optionally, a signer can also provide the nonce for the transaction to use.

Functions

get_signer_from_stronghold

Try to get signer details for an existing stronghold instance at the specific location

init_ephemeral_stronghold

Initialize the ephemeral stronghold client with a seed.

init_stronghold_from_path

Initialize a new stronghold instance from the provided snapshot path and passphrase

init_stronghold_from_seed

Initialize a new stronghold instance from the sr25519 mnemonic or raw seed

Type Definitions

KeyPair

Tidechain key-pair representation.

SubstrateSubxtError

An error that may contain some runtime error E

TidechainRuntimeApi

Tidechain runtime API generated from scale encoding file

TidefiKeyring

Tidefi keyring

TidefiPairSigner

Stronghold pair signer.

Derive Macros

Decode

Derive parity_scale_codec::Decode and for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.